Add test script + linting
Add example script (`create_disk_cache.py`) that creates a
`LlamaStaticDiskCache` from example snippets, then does some basic
sanity testing to make sure that the cache matches the prompt tokens as
expected.

The generation of an example prompt is slightly complex, because need to
generate the prompt with each context included, but not the user
question.